+++ packages/gamut-labs/src/GlobalHeader/__tests__/GlobalHeader-test.tsx
@@ -142,7 +142,7 @@ describe('GlobalHeader', () => {
 
     it('renders resourcesDropdown', () => {
       const { view } = renderView(anonHeaderProps);
-      view.getAllByText(resourcesDropdown.text);
+      view.getAllByText(resourcesDropdown().text);
     });
 
     it('renders communityDropdown', () => {
@@ -258,7 +258,7 @@ describe('GlobalHeader', () => {
 
       it('renders resourcesDropdown', () => {
         const { view } = renderView(freeHeaderProps);
-        view.getByText(resourcesDropdown.text);
+        view.getByText(resourcesDropdown().text);
       });
 
       it('renders communityDropdown', () => {
@@ -336,7 +336,7 @@ describe('GlobalHeader', () => {
 
       it('renders resourcesDropdown', () => {
         const { view } = renderView(proHeaderProps);
-        view.getByText(resourcesDropdown.text);
+        view.getByText(resourcesDropdown().text);
       });
 
       it('renders communityDropdown', () => {