+++ packages/gamut-labs/src/GlobalHeader/index.tsx
@@ -42,13 +42,13 @@ const getAppHeaderItems = (
     case 'anon':
       switch (props.variant) {
         case 'landing':
-          return anonLandingHeaderItems(props.hidePricing);
+          return anonLandingHeaderItems(props.hidePricing, props.user);
         case 'login':
-          return anonLoginHeaderItems(props.hidePricing);
+          return anonLoginHeaderItems(props.hidePricing, props.user);
         case 'signup':
-          return anonSignupHeaderItems(props.hidePricing);
+          return anonSignupHeaderItems(props.hidePricing, props.user);
         default:
-          return anonDefaultHeaderItems(props.hidePricing);
+          return anonDefaultHeaderItems(props.hidePricing, props.user);
       }
     case 'free':
       return freeHeaderItems(
@@ -70,13 +70,13 @@ const getMobileAppHeaderItems = (
     case 'anon':
       switch (props.variant) {
         case 'landing':
-          return anonLandingMobileHeaderItems(props.hidePricing);
+          return anonLandingMobileHeaderItems(props.hidePricing, props.user);
         case 'login':
-          return anonLoginMobileHeaderItems(props.hidePricing);
+          return anonLoginMobileHeaderItems(props.hidePricing, props.user);
         case 'signup':
-          return anonSignupMobileHeaderItems(props.hidePricing);
+          return anonSignupMobileHeaderItems(props.hidePricing, props.user);
         default:
-          return anonDefaultMobileHeaderItems(props.hidePricing);
+          return anonDefaultMobileHeaderItems(props.hidePricing, props.user);
       }
     case 'free':
       return freeMobileHeaderItems(props.user, props.hidePricing);