+++ packages/gamut-labs/src/Notifications/__tests__/useHeaderNotifications-test.tsx
@@ -62,7 +62,7 @@ describe('useHeaderNotifications', () => {
     );
 
     act(() => {
-      userEvent.click(buttonView.getByRole('button'));
+      userEvent.click(buttonView.getByRole('menuitem'));
     });
 
     const paneView = render(